schema change to support task archiving
The run_stream_fileset_assoc table in the Tier0 schema
holds a foreign key reference to the wmbs_fileset id
for the top level filesets for repack, express and
promptreco. When the TaskArchiver tries to delete
the subscription on that fileset, it also tries to
delete the fileset, which fails due to a foreign
key error.

I modified the Tier0 schema to cascade delete the
row in run_stream_fileset_assoc when the corresponding
fileset is deleted.